Move QuickLoginView to its own file
