+++ Tireless/Auth/View/QuickLoginView.swift
@@ -1,0 +1,42 @@
+//
+//  QuickLoginView.swift
+//  Tireless
+//
+//  Created by Liam on 2023/11/17.
+//
+
+import SwiftUI
+
+struct QuickLoginView: View {
+    
+    let width: CGFloat
+    
+    var body: some View {
+        Text("- Or sign up with -")
+            .font(.system(.body))
+            .foregroundColor(.gray)
+            .padding(.bottom, 20)
+        
+        HStack(spacing: 15) {
+            let buttonWidth = (width - 40) / 3
+            let buttonHeight = buttonWidth * 0.6
+            IconButton(iconImage: "star.fill", size: CGSize(width: buttonWidth, height: buttonHeight)) {
+                print("tap button 1")
+            }
+            IconButton(iconImage: "apple.logo", size: CGSize(width: buttonWidth, height: buttonHeight)) {
+                print("tap button 2")
+            }
+            IconButton(iconImage: "star", size: CGSize(width: buttonWidth, height: buttonHeight)) {
+                print("tap button 3")
+            }
+        }
+        .padding(.bottom, 30)
+        
+    }
+}
+
+struct QuickLoginView_Previews: PreviewProvider {
+    static var previews: some View {
+        QuickLoginView(width: 300)
+    }
+}
